cerebro-activa/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform>=1.38.0
python-multipart
gunicorn
